add hint when trying to export hidden notes and rename component accordingly
