--- src/components/SettingsDialog/ExportBoard/ExportHintHiddenColumns/index.ts
@@ -1,1 +1,0 @@
-export * from "./ExportHintHiddenColumns";
