+++ src/components/SettingsDialog/ExportBoard/ExportHintHiddenContent/ExportHintHiddenContent.scss
@@ -14,6 +14,7 @@ $list-style-type: "– ";
   grid-gap: $spacing--xs $spacing--base;
   margin-top: $spacing--lg;
   padding: 0 $spacing--base;
+  line-height: $line-height--medium;
 }
 
 .hint-hidden-columns__info-icon {
