+++ src/components/SettingsDialog/ExportBoard/index.ts
@@ -1,2 +1,1 @@
 export * from "./ExportBoard";
-export * from "./ExportHintHiddenColumns";
